Add list method to stores
